+++ bindings/ruby/spec/commerce_spec.rb
@@ -317,6 +317,6 @@ end
 
 RSpec.describe StateSet do
   it 'has a version number' do
-    expect(StateSet::VERSION).to eq('1.23.2')
+    expect(StateSet::VERSION).to eq('1.23.3')
   end
 end